+++ server/static/cal-heatmap-master/cal-heatmap.js
@@ -1,4 +1,4 @@
-/*! cal-heatmap v3.5.4 (Mon Aug 24 2015 10:02:24)
+/*! cal-heatmap v3.6.2 (Mon Oct 10 2016 01:36:20)
  *  ---------------------------------------------
  *  Cal-Heatmap is a javascript module to create calendar heatmap to visualize time series data
  *  https://github.com/wa0x6e/cal-heatmap
@@ -477,7 +477,7 @@ var CalHeatMap = function() {
 			extractUnit: function(d) {
 				var dt = new Date(d.getFullYear(), d.getMonth(), d.getDate());
 				// According to ISO-8601, week number computation are based on week starting on Monday
-				var weekDay = dt.getDay()-1;
+				var weekDay = dt.getDay() - (self.options.weekStartOnMonday ? 1 : 0);
 				if (weekDay < 0) {
 					weekDay = 6;
 				}
@@ -1423,6 +1423,7 @@ CalHeatMap.prototype = {
 
 				var htmlClass = parent.getHighlightClassName(d.t).trim().split(" ");
 				var pastDate = parent.dateIsLessThan(d.t, new Date());
+        var sameDate = parent.dateIsEqual(d.t, new Date());
 
 				if (parent.legendScale === null ||
 					(d.v === null && (options.hasOwnProperty("considerMissingDataAsZero") && !options.considerMissingDataAsZero) &&!options.legendColors.hasOwnProperty("base"))
@@ -1430,9 +1431,11 @@ CalHeatMap.prototype = {
 					htmlClass.push("graph-rect");
 				}
 
-				if (!pastDate && htmlClass.indexOf("now") === -1) {
-					htmlClass.push("future");
-				}
+        if (sameDate) {
+          htmlClass.push("now");
+        } else if (!pastDate) {
+          htmlClass.push("future");
+        }
 
 				if (d.v !== null) {
 					htmlClass.push(parent.Legend.getClass(d.v, (parent.legendScale === null)));
@@ -1875,8 +1878,8 @@ CalHeatMap.prototype = {
 
 		if (this.options.highlight.length > 0) {
 			for (var i in this.options.highlight) {
-				if (this.options.highlight[i] instanceof Date && this.dateIsEqual(this.options.highlight[i], d)) {
-					return " highlight" + (this.isNow(this.options.highlight[i]) ? " now": "");
+				if (this.dateIsEqual(this.options.highlight[i], d)) {
+					return this.isNow(this.options.highlight[i]) ? " highlight-now": " highlight";
 				}
 			}
 		}
@@ -1908,6 +1911,14 @@ CalHeatMap.prototype = {
 	/* jshint maxcomplexity: false */
 	dateIsEqual: function(dateA, dateB) {
 		"use strict";
+
+		if(!(dateA instanceof Date)) {
+			dateA = new Date(dateA);
+		}
+
+		if (!(dateB instanceof Date)) {
+			dateB = new Date(dateB);
+		}
 
 		switch(this.options.subDomain) {
 		case "x_min":
@@ -2923,8 +2934,10 @@ CalHeatMap.prototype = {
 			".graph-rect": {},
 			"rect.highlight": {},
 			"rect.now": {},
+			"rect.highlight-now": {},
 			"text.highlight": {},
 			"text.now": {},
+			"text.highlight-now": {},
 			".domain-background": {},
 			".graph-label": {},
 			".subdomain-text": {},